refactor: Refactor and restructure chart components
This commit refactors the chart components for better organization and clarity.

Key changes include:
- Replaced the `ChartDescription` composable with a `BasicText` implementation directly within `LineChart`.
- Restructured the chart component files into more specific sub-packages (`axis`, `grid`, `line`, `tooltip`).
- Removed the unused `title` parameter from `ChartContent` and `drawQuarticLineWithShadow`.
- Updated the default description text style to use `Theme.typography.label.medium`.
- Set a default `valueSuffix` in `SpendingTrendChart`.
