+++ design_system/src/commonMain/kotlin/com/moneyplusplus/design_system/chart/LineChart.kt
@@ -1,6 +1,7 @@
-package com.moneyplusplus.design_system.chart.components
+package com.moneyplusplus.design_system.chart
 
 import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.text.BasicText
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.remember
@@ -8,9 +9,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
+import com.moneyplusplus.design_system.chart.components.ChartContent
 import com.moneyplusplus.design_system.chart.models.ChartColors
 import com.moneyplusplus.design_system.chart.utils.ChartDefaultValues
-
+import com.moneyplusplus.design_system.theme.theme.Theme
 
 @Composable
 fun LineChart(
@@ -23,7 +25,7 @@ fun LineChart(
     barWidthPx: Dp = ChartDefaultValues.backgroundLineWidth,
     animateChart: Boolean = ChartDefaultValues.ANIMATED_CHART,
     showGridWithSpacer: Boolean = ChartDefaultValues.SHOW_BACKGROUND_WITH_SPACER,
-    descriptionStyle: TextStyle = ChartDefaultValues.descriptionDefaultStyle,
+    descriptionStyle: TextStyle = Theme.typography.label.medium,
     yAxisStyle: TextStyle = ChartDefaultValues.axesStyle,
     xAxisStyle: TextStyle = ChartDefaultValues.axesStyle,
     chartRatio: Float = ChartDefaultValues.chartRatio,
@@ -34,22 +36,24 @@ fun LineChart(
     Box(modifier.wrapContentHeight()) {
         Column() {
             // Title
-            Box(Modifier.padding(bottom = 16.dp)) {
-                ChartDescription(
-                    chartName = title,
-                    descriptionStyle = descriptionStyle,
-                )
+            if (title.isNotEmpty()) {
+                Box(Modifier.padding(bottom = 16.dp)) {
+                    BasicText(
+                        text = title,
+                        style = descriptionStyle.copy(color = chartColors.axisLabelColor)
+                    )
+                }
             }
 
             ChartContent(
                 modifier = if (chartRatio == 0f) Modifier.wrapContentSize()
                 else Modifier.aspectRatio(chartRatio).fillMaxSize(),
                 data = data,
-                title = title,
+                // title parameter removed
                 chartColors = chartColors,
                 valueSuffix = valueSuffix,
-                tooltipBackgroundColor = chartColors.tooltipBackground,
-                tooltipTextColor = chartColors.tooltipTextColor,
+                tooltipBackgroundColor = chartColors.tooltipBackground, 
+                tooltipTextColor = chartColors.tooltipTextColor, 
                 xAxisData = xAxisData,
                 barWidthPx = barWidthPx,
                 animateChart = animateChart,
