+++ design_system/src/commonMain/kotlin/com/moneyplusplus/design_system/chart/components/axis/YAxisDrawing.kt
@@ -1,4 +1,4 @@
-package com.moneyplusplus.design_system.chart.components
+package com.moneyplusplus.design_system.chart.components.axis
 
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.drawscope.DrawScope
