+++ design_system/src/commonMain/kotlin/com/moneyplusplus/design_system/chart/utils/ChartDefaultValues.kt
@@ -2,7 +2,6 @@ package com.moneyplusplus.design_system.chart.utils
 
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 
@@ -12,12 +11,6 @@ internal object ChartDefaultValues {
     val backgroundLineWidth = 1.dp
     const val SHOW_BACKGROUND_WITH_SPACER = true
     const val chartRatio = 0f
-    val descriptionDefaultStyle = TextStyle(
-        color = Color.Black,
-        fontSize = 14.sp,
-        fontWeight = FontWeight.W400
-    )
-
     val axesStyle = TextStyle(
         fontSize = 12.sp,
         color = Color.Gray,
